hike
This C++ code defines a system for managing hiking trails, encapsulating essential trail attributes within the Traseu class, such as distance, duration, points of interest, and map details. Additionally, it introduces a specialized TraseuMontan class to represent mountain trails, extending the functionality of the base class with altitude information. Through method overrides and operator overloading, it enables the manipulation and comparison of trail instances. The main function showcases the instantiation of various trail objects and demonstrates their attributes, emphasizing polymorphic behavior through the virtual afisare() method. Moreover, it incorporates a template function suma() to calculate the sum of durations for any given trail types, highlighting the flexibility and reusability of the code.